requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
docopt==0.6.2
h11==0.13.0
idna==3.3
lxml==4.8.0
outcome==1.1.0
pipreqs==0.4.11
pycparser==2.21
pyOpenSSL==22.0.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
webdriver-manager==3.5.4
wincertstore==0.2
wsproto==1.0.0
yarg==0.1.9
